+++ src/botvinnik/plugins/Debian.hpp
@@ -58,6 +58,15 @@ class Debian final: public DeactivatablePlugin
      * \return Returns a short, one line help text for the command.
      */
     std::string helpOneLine(const std::string_view& command) const override;
+
+
+    /** \brief Gets a longer help text for a command.
+     *
+     * \param command  name of the command where help is requested
+     * \param prefix   the prefix for bot commands, e. g. "!"
+     * \return Returns a Message containing a longer help text for the command.
+     */
+    Message helpExtended(const std::string_view& command, const std::string_view& prefix) const override;
   private:
     /** Search results: packages names and (possibly) versions */
     struct Packages