+++ src/botvinnik/plugins/Fortune.cpp
@@ -97,4 +97,19 @@ std::string Fortune::helpOneLine(const std::string_view& command) const
   return std::string();
 }
 
+Message Fortune::helpExtended(const std::string_view& command,
+             [[maybe_unused]] const std::string_view& prefix) const
+{
+  if (command == "fortune")
+  {
+    return Message("displays a random epigram / fortune cookie / quote");
+  }
+  if (command == "fortunes")
+  {
+    return Message("displays a random epigram / fortune cookie / quote (alias for the fortune command)");
+  }
+
+  return Message();
+}
+
 } // namespace