+++ src/botvinnik/plugins/corona/Corona.hpp
@@ -65,6 +65,15 @@ class Corona final: public DeactivatablePlugin
      * \return Returns a short, one line help text for the command.
      */
     std::string helpOneLine(const std::string_view& command) const override;
+
+
+    /** \brief Gets a longer help text for a command.
+     *
+     * \param command  name of the command where help is requested
+     * \param prefix   the prefix for bot commands, e. g. "!"
+     * \return Returns a Message containing a longer help text for the command.
+     */
+    Message helpExtended(const std::string_view& command, const std::string_view& prefix) const override;
   private:
     std::optional<std::pair<std::string, std::chrono::steady_clock::time_point> > dbLocation; /**< location and update time of the database file */
     Matrix& theMatrix; /**< reference to the Matrix client */