+++ tests/plugins/core/Help.cpp
@@ -1,7 +1,7 @@
 /*
  -------------------------------------------------------------------------------
     This file is part of the test suite for botvinnik.
-    Copyright (C) 2020, 2022  Dirk Stolle
+    Copyright (C) 2020, 2022, 2023  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -52,6 +52,23 @@ TEST_CASE("plugin Help")
   SECTION("one line help for non-existent command returns no text")
   {
     REQUIRE( plugin.helpOneLine("plonk").empty() );
+  }
+
+  SECTION("extended help")
+  {
+    for (const auto& cmd : commands)
+    {
+      // Help text must not be empty.
+      const Message msg = plugin.helpExtended(cmd, "!");
+      REQUIRE_FALSE( msg.body.empty() );
+    }
+  }
+
+  SECTION("extended help for non-existent command returns no text")
+  {
+    const Message msg = plugin.helpExtended("plonk", "!");
+    REQUIRE( msg.body.empty() );
+    REQUIRE( msg.formatted_body.empty() );
   }
 
   SECTION("command handlers must return text")